% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Wnt_denovo_genesets.R
\name{Get_Wnt_denovo_genesets}
\alias{Get_Wnt_denovo_genesets}
\title{Get_Wnt_denovo_genesets}
\usage{
Get_Wnt_denovo_genesets(
  file_paths,
  expression_accession_vector,
  group_HL,
  gene_difference_method = "limma",
  alternative = "two.sided",
  p_combine_method = "fisher",
  threshold_or_rank = "rank",
  top_genes = 500,
  gene_Pfilter = 0.05,
  gene_FCfilter = 1,
  export_file = F
)
}
\arguments{
\item{file_paths}{Working directory path.}

\item{expression_accession_vector}{A vector of all sub-dataset names to be used for de novo Wnt pathway-related gene set identification.}

\item{group_HL}{A file required for each sub-dataset to distinguish between high (H) and low (L) Wnt signaling activity groups. File column 1: Accession, column 2: Group1, column 3: Group2, column 4: Group1_Status, column 5: Group2_Status.}

\item{gene_difference_method}{Method for identifying differentially expressed genes associated with Wnt pathway activity: "limma", "t_test", "wilcox_test".}

\item{alternative}{When using t.test/wilcox.test, select "two.sided", "less", or "greater" to assess Wnt activity-associated expression changes.}

\item{p_combine_method}{Method for integrating p-values across multiple sub-datasets within the same Wnt-related study, options include "fisher", "z.transform", "logit", "ctt".}

\item{threshold_or_rank}{Strategy for selecting Wnt pathway-associated differentially expressed genes: "rank" for ranking by p-value, or "threshold" for applying significance cutoffs.}

\item{top_genes}{If "rank" is selected for threshold_or_rank, the top-ranked genes associated with differential Wnt pathway activity will be retained (default = 500).}

\item{gene_Pfilter}{If "threshold" is selected for threshold_or_rank, this defines the p-value cutoff for Wnt-related differential expression (default = 0.05).}

\item{gene_FCfilter}{If "threshold" is selected for threshold_or_rank, this defines the log2 fold change threshold for Wnt pathway activity-associated differential expression (default = 1).}

\item{export_file}{Logical. If TRUE, write the de novo gene sets to tab-delimited text files in \code{file_paths}.}
}
\value{
A list of de novo gene sets representing Wnt pathway activation (up-regulated genes) and inhibition (down-regulated genes).
}
\description{
To obtain de novo Wnt signaling pathway-related gene sets from multiple bulk RNA-Seq datasets,
differentially expressed genes associated with Wnt pathway activity are first identified via limma/t.test/wilcox.test,
then up-regulated and down-regulated differentially expressed genes are designated as two de novo gene sets representing Wnt pathway activation and inhibition, respectively.
}
\author{
Dingkang Zhao
}
