% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Wnt_genesets.R
\name{Merge_Wnt_genesets}
\alias{Merge_Wnt_genesets}
\title{Merge_Wnt_genesets}
\usage{
Merge_Wnt_genesets(
  file_paths,
  activation_geneset,
  inhibition_geneset,
  delete_GN = T,
  delete_time = c("before", "after"),
  min_GN = 5,
  integration_method = c("Jaccard", "Sorensen-Dice", "Hub-Promoted", "Hub-Depressed"),
  de_redundant_basis = c("igraph_component", "agglomeration_ward.D",
    "agglomeration_ward.D2", "agglomeration_single", "agglomeration_complete",
    "agglomeration_average", "agglomeration_mcquitty", "agglomeration_median",
    "agglomeration_centroid"),
  similarity = 0.3,
  export_file = F
)
}
\arguments{
\item{file_paths}{Working directory path.}

\item{activation_geneset}{Wnt pathway activation gene sets to be merged. (Place in the working directory)}

\item{inhibition_geneset}{Wnt pathway inhibition gene sets to be merged. (Place in the working directory)}

\item{delete_GN}{TRUE: Remove gene sets containing insufficient numbers of Wnt-related genes. FALSE: Retain all gene sets. Default parameter is "TRUE".}

\item{min_GN}{Gene sets with fewer genes than min_GN will be removed from the Wnt pathway collection. Default parameter is 5.}

\item{similarity}{When the Jaccard index similarity between two Wnt-related gene sets exceeds this threshold, the gene sets will be integrated. Default parameter is 0.3.}
}
\description{
Removal of gene sets containing few genes after purification, and integration of biologically similar Wnt-related gene sets through construction of Jaccard index networks.
}
\author{
Dingkang Zhao
}
